requirements.txt
slack-bolt==1.18.0
python-dotenv==1.0.0
openai==1.12.0
google-cloud-aiplatform==1.42.1
google-generativeai==0.3.2